+++ src/main/java/consumer/GasConsumer.java
@@ -6,7 +6,7 @@ import event.LeakEvent;
 public interface GasConsumer extends Consumer {
     int GAS_COST = 0;   // cost for unit of gas TODO - move to Constants and give value
 
-    int consumeGas();
+    double consumeGas();
 
     default void leak() {
         if (this instanceof Device device) {    // TODO - can we remove instanceof?
