+++ src/main/java/consumer/WaterConsumer.java
@@ -6,7 +6,7 @@ import event.FloodEvent;
 public interface WaterConsumer extends Consumer {
     int WATER_COST = 0;   // cost for unit of water TODO - move to Constants and give value
 
-    int consumeWater();
+    double consumeWater();
 
     default void flood() {
         if (this instanceof Device device) {    // TODO - can we remove instanceof?
